Use integers only instead of longs for serial commands received by Arduino, change serial scaling factor to 100, add limits to target values in GUI to help enforce precision limits
System faced memory allocation issues, it seems, when longs were used. It was weird and hard to debug. Sticking with integers was better, though you have to sacrifice on precision. Changed serial scaling factor to 100, which gives max values of 327.67 (2 decimals precision) which can be sent to the Arduino. Plenty, considering the types of values that the GUI will send.